+++ packages/arb-token-bridge-ui/src/pages/api/withdrawals.ts
@@ -2,26 +2,7 @@ import { NextApiRequest, NextApiResponse } from 'next'
 import { gql } from '@apollo/client'
 
 import { getL2SubgraphClient } from '../../api-utils/ServerSubgraphUtils'
-
-export type FetchWithdrawalsFromSubgraphResult = {
-  id: string
-  type: 'EthWithdrawal' | 'TokenWithdrawal'
-  sender: string
-  receiver: string
-  ethValue: string
-  l1Token?: {
-    id: string
-  }
-  tokenAmount: string
-  isClassic: boolean
-  l2BlockTimestamp: string
-  l2TxHash: string
-  l2BlockNum: string
-  direction: 'withdrawal'
-  source: 'subgraph'
-  parentChainId: number
-  childChainId: number
-}
+import { FetchWithdrawalsFromSubgraphResult } from '../../util/withdrawals/fetchWithdrawalsFromSubgraph'
 
 // Extending the standard NextJs request with Withdrawal-params
 type NextApiRequestWithWithdrawalParams = NextApiRequest & {
